refactor: move general components to separate directory
