+++ components/awards.js
@@ -1,8 +1,8 @@
 import { html } from '@rbardini/html'
 import markdown from '../utils/markdown.js'
-import Article from './article.js'
-import DateTime from './date-time.js'
-import Section from './section.js'
+import Article from './general/article.js'
+import DateTime from './general/date-time.js'
+import Section from './general/section.js'
 
 /**
  * @param {import('../schema.d.ts').ResumeSchema['awards']} awards
